remove JSONL read_timeline fallback; SQLite-only timeline path (#67)
